requirements.txt
Flask==3.1.1
flask-sock==0.7.0
paho-mqtt==2.1.0
tk==0.1.0
webcolors==24.11.1